refactor: move SubjectPhasePicker to SharedComponents
