+++ src/components/pages/LandingPages/CurriculumLandingHero.tsx
@@ -5,7 +5,7 @@ import Flex from "@/components/SharedComponents/Flex";
 import Box from "@/components/SharedComponents/Box";
 import SubjectPhasePicker, {
   SubjectPhasePickerData,
-} from "@/components/SubjectPhasePicker/SubjectPhasePicker";
+} from "@/components/SharedComponents/SubjectPhasePicker/SubjectPhasePicker";
 import Illustration from "@/components/SharedComponents/Illustration/Illustration";
 import { getSizes } from "@/components/SharedComponents/CMSImage/getSizes";
 import Cover from "@/components/SharedComponents/Cover/Cover";
